+++ packages/vex-sdk-mock/src/device.rs
@@ -33,7 +33,7 @@ pub extern "C" fn vexDeviceGetTimestamp(device: V5_DeviceT) -> u32 {
     Default::default()
 }
 #[unsafe(no_mangle)]
-pub extern "C" fn vexDeviceGenericValueGet(device: V5_DeviceT) -> c_double {
+pub extern "C" fn vexDeviceGenericValueGet(device: V5_DeviceT) -> i32 {
     Default::default()
 }
 #[unsafe(no_mangle)]
